Re-organize tessellation and rasterization-related code
This is a refactor/re-organization of the tessellation and
rasterization-related code, to help with larger modeling of the library;
we now have a tessellation domain (tess) and a rasterization domain
(raster).

This will help with further re-factoring down the line.

Additionally, as part of this, each rasterizer has been separated into
individual source files, one each for non-AA, SSAA, and MSAA.