Add hairline stroke support
This commit adds hairline strokes.

These strokes are drawn using one of two algorithms: Bresenham (non-AA)
or Xiaolin Wu (any other AA mode). They are designed to render strokes
at the minimum whole display unit width; at the current stage of
development with the CPU renderer, this is one pixel wide.

Note that while hairline stroking is likely faster than using the
traditional stroking mode, performance is not necessarily the goal, but
rather better accuracy for these scenarios. At their core, Bresenham and
Wu draw one pixel at a time, and as such incur the performance penalties
of doing so. Horizontal hairlines are, however, reduced to a memset.

Dashed lines are currently not implemented, but will be in the future.

Note as well that YMMV when drawing arcs due to the inherent nature of
how we reduce arcs to splines, and splines down to polylines. This means
that better results may be had through rolling of a native midpoint or
other dedicated circle algorithm, although this is out of the scope of
this work.